+++ src/main/java/com/example/demo/repo/DoctorRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entity.Doctor;
 
-public interface DoctorRepo extends JpaRepository<Doctor, Long>
+public interface DoctorRepo extends JpaRepository<Doctor, Integer>
 {
