+++ src/main/java/com/example/demo/repo/PatientRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entity.Patient;
 
-public interface PatientRepo extends JpaRepository<Patient, Long>
+public interface PatientRepo extends JpaRepository<Patient, Integer>
 {
 
 }
