+++ src/main/java/com/example/demo/service/DoctorService.java
@@ -7,8 +7,8 @@ import com.example.demo.entity.Doctor;
 public interface DoctorService 
 {
 	public List<Doctor> getallDoctor();
-    public Doctor getDoctorById(Long id);
+    public Doctor getDoctorById(int id);
 	public Doctor createDoctor(Doctor doctor);
-    public void deleteDoctor(Long id);
+    public void deleteDoctor(int id);
 
 }
